Generate map colors as vec
Since province ids are generally sequential starting at 1, we can just have a vec instead of a hashmap. The unused indices are still cheaper than the cost of a hashmap. Note that this could change if there is ever a map (probably from a mod) that uses some very large province id.