Partial data sanitization
There are a lot of anomalies in the historical data. Most of this seems to come from tag formation, where the formed tag is given 'control' of provinces before it is formed. Tags formations also don't trigger province ownership events, so we have to do that conversion ourselves. This commit cleans up those cases.

The other 10% of province control anomalies have an unclear origin and seem to be missing control change events. I'm not sure what causes them, so that's a TODO.