Optimize stats memory usage
In the browser it's fine, but the bot has less than 256mb on the legacy free tier plan I'm using and a not particularly fast cpu. By manually dropping as early as possible and otherwise reducing memory footprint, we can avoid using slow swap memory.
It's still very slow compared to running on a proper machine, so we may want a different solution, though I still want to see if I can make this work only using free resources.