Color EU5 unownable locations
Unownable locations (uninhabitable/impassible terrain and desert corridors) get a country's color if all adjacent ownable locations are owned by that country. In the current state of the game (1.0.7), the game itself does not actually do this properly in all cases, but this has been ackowledged as a bug so we will use this intended method for choosing the color of unownable locations.